pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "hatch-vcs"]

[project]
name = "wsifir"
description = "things"
authors = [
    {name = "Olamide Olanrewaju"},
]
maintainers = [
    {name = "Olamide Olanrewaju", email = "mide.olanrewaju@queensu.ca"},
    {name = "Benjamin Green", email = "bgreen42@jh.edu"},
]
dynamic = ["version"]
requires-python = ">=3.11"

license = {file = "LICENSE"}
readme = "README.md"
dependencies = [
    "antspyx",
    "dask",
    "ipywidgets",
    "matplotlib",
    "numpy",
    "pandas",
    "pyimagej",
    "scikit-learn",
    "scikit-image",
    "scipy",
    "sidus @ git+https://github.com/AstroPathJHU/sidus.git@main",
    "simpleitk",
]

[project.optional-dependencies]

test = [
    "black",
    "flake8",
    "pylint",
    "pytest",
    "pylint-pytest",
    "pytest-datafiles",
    "pre-commit",
]

jupyter = [
    "ipykernel",
    "ipython>=8.6.0",
    "jupyter",
]

dev = [
    "wsifir[test,jupyter]",
]

[tool.black]
line-length = 80

[tool.hatch.build.targets.wheel]
packages = ['src/wsifir']

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "_version.py"

[tool.hatch.metadata]
allow-direct-references = true
